Revert "Adds sign up form"
This reverts commit 9e92e9692b094b62b688da4e07e254b7d9637b7d.
