+++ src/_components/customs/Header/UserMenu.tsx
@@ -15,15 +15,14 @@ export function UserMenu({ user }) {
         <NavigationMenuList>
           <NavigationMenuItem key={user?.name}>
             <NavigationMenuTrigger>
-              {user ? user?.name ?? user.email : 'Login / Signup'}
+              {user ? user?.name ?? user.email : 'Login'}
             </NavigationMenuTrigger>
             <NavigationMenuContent>
               <ul className="grid gap-2  p-6 md:w-[120px] ]">
                 <ListItem
-                  title={user?.name ? 'Profile' : 'Login'}
+                  title={user?.name ? 'Profile' : 'LogIn'}
                   href={user ? '/profile' : '/login'}
                 />
-                {!user && <ListItem title="Signup" href="/signup" />}
                 {user?.roles?.includes('admin') && <ListItem title="Dash" href="/admin" />}
                 {user && <ListItem title="Logout" href="/logout" />}
               </ul>
