+++ src/app/login/LoginForm/index.tsx
@@ -20,67 +20,57 @@ import {
   FormMessage,
 } from '@/components/ui/form'
 import { Input } from '@/components/ui/input'
+import { User } from '@/payload/payload-types'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useRouter } from 'next/navigation'
-import Router from 'next/router'
+import { userAgent } from 'next/server'
 import { useState } from 'react'
 import { RegisterOptions, useForm, UseFormRegisterReturn } from 'react-hook-form'
 import * as z from 'zod'
 
 const formSchema = z.object({
-  name: z.string().min(1, { message: 'Please enter your name.' }),
   email: z.string().email({
     message: 'Please enter a valid email address.',
   }),
-  password: z.string().min(3, {
+  password: z.string().min(8, {
     message: 'Password must be at least 8 characters long.',
   }),
 })
 
-export default function SignUpForm() {
-  const [error, setError] = useState('')
-  const [formMessage, setFormMessage] = useState<null | string>(null)
-  const { user, setUser } = useAuth()
+export default function MyLogin() {
   const router = useRouter()
+  const [error, setError] = useState('')
+  const { setUser } = useAuth()
 
   const form = useForm({
     resolver: zodResolver(formSchema),
     defaultValues: {
       email: '',
       password: '',
-      name: '',
     },
   })
 
   async function onSubmit(values) {
     try {
-      const req = await fetch(`${process.env.NEXT_PUBLIC_SERVER_URL}/api/users`, {
+      const response = await fetch(`${process.env.NEXT_PUBLIC_SERVER_URL}/api/users/login`, {
         method: 'POST',
-        credentials: 'include',
-        headers: {
-          'Content-Type': 'application/json',
-        },
+        headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify(values),
       })
 
-      if (!req.ok) {
-        const errorData = await req.json()
-        throw new Error(errorData.message || `HTTP error! status: ${req.status}`)
-      }
-
-      const data = await req
-        .json()
-        .then(data => setFormMessage(`${data.message}`))
-        .then(() => {
-          setTimeout(() => {
-            {
-              window.location.reload()
-            }
-          }, 3000)
+      if (response.ok) {
+        await response.json().then(data => {
+          const user: User = data.user
+          setUser(user)
         })
-    } catch (err) {
-      console.log('error', err)
-      setError(err.message)
+
+        router.push('/home') // Redirect to dashboard on successful login
+      } else {
+        const data = await response.json()
+        setError(data.message || 'Login failed')
+      }
+    } catch (error) {
+      setError('An error occurred. Please try again.')
     }
   }
 
@@ -88,40 +78,18 @@ export default function SignUpForm() {
     <div className="flex items-center justify-center h-100vh bg-gray-100">
       <Card className="w-[350px]">
         <CardHeader>
-          <CardTitle>SignUp</CardTitle>
-          <CardDescription>Enter your credentials create your account.</CardDescription>
+          <CardTitle>Login</CardTitle>
+          <CardDescription>Enter your credentials to access your account.</CardDescription>
         </CardHeader>
         <CardContent>
           <Form {...form}>
-            <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
-              <FormField
-                control={form.control}
-                name="name"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Name</FormLabel>
-                    <FormControl>
-                      <Input
-                        register={function <TFieldName extends string = string>(
-                          name: TFieldName,
-                          options?: RegisterOptions<any, TFieldName>,
-                        ): UseFormRegisterReturn<TFieldName> {
-                          throw new Error('Function not implemented.')
-                        }}
-                        placeholder="Enter your Name"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
+            <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
               <FormField
                 control={form.control}
                 name="email"
                 render={({ field }) => (
                   <FormItem>
-                    <FormLabel>Email </FormLabel>
+                    <FormLabel>Email</FormLabel>
                     <FormControl>
                       <Input
                         register={function <TFieldName extends string = string>(
@@ -130,7 +98,7 @@ export default function SignUpForm() {
                         ): UseFormRegisterReturn<TFieldName> {
                           throw new Error('Function not implemented.')
                         }}
-                        placeholder={`Enter email.`}
+                        placeholder="Enter your email"
                         {...field}
                       />
                     </FormControl>
@@ -166,18 +134,17 @@ export default function SignUpForm() {
                   <AlertDescription>{error}</AlertDescription>
                 </Alert>
               )}
-              {!error && formMessage && <Alert variant="default">{formMessage}</Alert>}
               <Button type="submit" className="w-full">
-                Sign up
+                Login
               </Button>
             </form>
           </Form>
         </CardContent>
         <CardFooter className="flex justify-center">
           <p className="text-sm text-gray-600">
-            Already have an account?
-            <a href="/login" className="text-blue-600 hover:underline">
-              {` Login`}
+            Don't have an account?{' '}
+            <a href="/signup" className="text-blue-600 hover:underline">
+              Sign up
             </a>
           </p>
         </CardFooter>
